chore: move help tooltip
