+++ apps/web/src/components/Composer/Actions/AccessSettings/index.tsx
@@ -1,4 +1,3 @@
-import HelpTooltip from '@components/UI/HelpTooltip';
 import { Modal } from '@components/UI/Modal';
 import { Tooltip } from '@components/UI/Tooltip';
 import { LockClosedIcon } from '@heroicons/react/outline';
@@ -8,6 +7,7 @@ import { motion } from 'framer-motion';
 import type { FC } from 'react';
 import { useState } from 'react';
 import { useAccessSettingsStore } from 'src/store/access-settings';
+import { HelpTooltip } from 'ui';
 
 import BasicSettings from './BasicSettings';
