+++ apps/web/src/components/Publication/Actions/Analytics/Stats.tsx
@@ -1,12 +1,11 @@
 import Loader from '@components/Shared/Loader';
-import HelpTooltip from '@components/UI/HelpTooltip';
 import { t } from '@lingui/macro';
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { ERROR_MESSAGE, SIMPLEANALYTICS_API } from 'data/constants';
 import type { Publication } from 'lens';
 import type { FC } from 'react';
-import { ErrorMessage } from 'ui';
+import { ErrorMessage, HelpTooltip } from 'ui';
 import humanize from 'utils/humanize';
 
 const Stat: FC<{ title: string; helper: string; stat: number }> = ({ title, helper, stat }) => (
