+++ packages/ui/index.ts
@@ -6,3 +6,4 @@ export { EmptyState } from './EmptyState';
 export { ErrorMessage } from './ErrorMessage';
 export { FieldError, Form, useZodForm } from './Form';
 export { GridItemEight, GridItemFour, GridLayout } from './GridLayout';
+export { default as HelpTooltip } from './HelpTooltip';
